# ddtrace/tracer: use math/rand/v2 for SpanID and TraceID